+++ src/ResourceManager/RecoveryServices/Commands.RecoveryServices.Backup/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Added policy filter to Get-AzureRmRecoveryServicesBackItem cmdlet. The command returns the list of backup items protected by the given policy id.
 
 ## Version 4.3.1
 * Updated all help files to include full parameter types and correct input/output types.