+++ src/ResourceManager/RecoveryServices/Commands.RecoveryServices.Backup/ChangeLog.md
@@ -18,8 +18,10 @@
         - Additional information about change #1
 -->
 ## Current Release
-* Updated all help files to include full parameter types and correct input/output types.
 * Added policy filter to Get-AzureRmRecoveryServicesBackItem cmdlet. The command returns the list of backup items protected by the given policy id.
+
+## Version 4.3.1
+* Updated all help files to include full parameter types and correct input/output types.
 
 ## Version 4.3.0
 * Fixed formatting of OutputType in help files